Extract federation functions into functions-federation.php
Move federation-related functions from functions.php to a dedicated file
for better maintainability:

- set_wp_object_state()
- get_wp_object_state()
- is_self_ping()
- add_to_outbox()
- follow()
- unfollow()

This reduces functions.php by ~220 lines and groups related functionality.
Also reorder includes to load constants.php before functions.php.
